The confirm slot becomes a hook of its own
One promise, one resolver ref, one ConfirmRequest — the whole `window.confirm`
replacement (#1082) reads the same from the outside and no longer sits in the
middle of the composer's load effects.
